add 'fonts' and touchup 'install' make targets
"make fonts" now installs the powerline inconsolada font at the expected
place.

"make install" now prereqs the "fonts" target, and is conditionalized to not
try to recreate the ~/.vimrc link if it already exists.

...though, right now, it only checks to see that ~/.vimrc exists, not that
it's a link to our vimrc.   We may want to rethink and refactor this to save
the original and link our vimrc if it's not already a link.