Include Interesting Remotes via `:Projects!` ... kinda
`:Projects` allows us to quickly scan project directories.  However,
we may be looking for something that does not yet reside on this
computer, and so `:Projects!` should support remote searches as well.

In reality, for speed (and API rate-limiting) considerations we're going
to want a local list of interesting git remotes.  When selected, the
given remote should then be cloned somewhere interesting.

This is the start of that: `:Projects!` works, but not cloning or
picking a place to clone to.  Nor is the code to hit up the GitHub API
to pull a list of clone URLs for all our repos as well as those we watch
(and maybe those of people we watch, too).